+++ src/indexOf.d.ts
@@ -1,0 +1,7 @@
+import type {ImmutableTree} from './types';
+
+export default function indexOf<T, K = T>(
+  tree: ImmutableTree<T> | null,
+  key: K,
+  cmp: (a: K, b: T) => number,
+): number;